Warnings

+++ Auximov/Auximov/Game.cpp
@@ -1,8 +1,8 @@
 #include "Game.h"
 
 Game::Game(sf::RenderWindow* app) : App(app) {
-	view1.setCenter(sf::Vector2f(App->getSize().x / 2, App->getSize().y / 2));
-	view1.setSize(sf::Vector2f(App->getSize().x, App->getSize().y));
+	view1.setCenter(sf::Vector2f((float)(App->getSize().x / 2), (float)(App->getSize().y / 2)));
+	view1.setSize(sf::Vector2f((float)App->getSize().x, (float)App->getSize().y));
 
 }
